[SSAP-20] removed the functionality to parse floating point numbers (too complicated)

+++ parser/parser/Expression.cpp
@@ -1,12 +1,12 @@
 #include "Expression.h"
 
 
-float Expression::parseExpression(std::string input) {
+int Expression::parseExpression(std::string input) {
 	input.erase(remove(input.begin(), input.end(), ' '), input.end()); // remove all whitespaces in expression
 	return expression(input);
 }
 
-float Expression::expression(std::string input, size_t offset) {
+int Expression::expression(std::string input, size_t offset) {
 	size_t plusLoc = input.find('+', offset);
 	size_t minusLoc = input.find('-', offset);
 	size_t openParentLoc = input.find('(', offset);
@@ -14,8 +14,8 @@ float Expression::expression(std::string input, size_t offset) {
 
 	if (plusLoc != std::string::npos && plusLoc < minusLoc) {
 		if ((plusLoc < openParentLoc || plusLoc > closeParentLoc)) { // if '+' is before '(' or after ')' and '+' is before '-'
-			float t = term(input.substr(0, plusLoc));
-			float e = expression(input.substr(plusLoc + 1, input.length() - plusLoc - 1));
+			int t = term(input.substr(0, plusLoc));
+			int e = expression(input.substr(plusLoc + 1, input.length() - plusLoc - 1));
 
 			std::cout << "Expression: " << t << " + " << e << "\n";
 
@@ -32,8 +32,8 @@ float Expression::expression(std::string input, size_t offset) {
 	}
 	if (minusLoc != std::string::npos) { // if there is no '+'
 		if (minusLoc < openParentLoc || minusLoc > closeParentLoc) { // if '-' is before '(' or after ')'
-			float t = term(input.substr(0, minusLoc));
-			float e = expression(input.substr(minusLoc + 1, input.length() - minusLoc - 1));
+			int t = term(input.substr(0, minusLoc));
+			int e = expression(input.substr(minusLoc + 1, input.length() - minusLoc - 1));
 
 			std::cout << "Expression: " << t << " - " << e << "\n";
 
@@ -54,12 +54,12 @@ float Expression::expression(std::string input, size_t offset) {
 
 }
 
-float Expression::term(std::string input, size_t offset) {
+int Expression::term(std::string input, size_t offset) {
 	size_t multiplyLoc = input.find('*', offset);
 
 	if (multiplyLoc != std::string::npos) {
-		float f = factor(input.substr(0, multiplyLoc));
-		float t = term(input.substr(multiplyLoc + 1, input.length() - multiplyLoc - 1));
+		int f = factor(input.substr(0, multiplyLoc));
+		int t = term(input.substr(multiplyLoc + 1, input.length() - multiplyLoc - 1));
 
 		std::cout << "Term: " << f << " * " << t << "\n";
 
@@ -69,8 +69,8 @@ float Expression::term(std::string input, size_t offset) {
 		size_t divideLoc = input.find('/', offset);
 
 		if (divideLoc != std::string::npos) {
-			float f = factor(input.substr(0, divideLoc));
-			float t = term(input.substr(divideLoc + 1, input.length() - divideLoc - 1));
+			int f = factor(input.substr(0, divideLoc));
+			int t = term(input.substr(divideLoc + 1, input.length() - divideLoc - 1));
 
 			std::cout << "Term: " << f << " / " << t << "\n";
 
@@ -83,7 +83,7 @@ float Expression::term(std::string input, size_t offset) {
 	}
 }
 
-float Expression::factor(std::string input, size_t offset) {
+int Expression::factor(std::string input, size_t offset) {
 	size_t openParentLoc = input.find('(', offset);
 
 	if (openParentLoc != std::string::npos) {
@@ -101,8 +101,8 @@ float Expression::factor(std::string input, size_t offset) {
 	}
 }
 
-float Expression::parseNumber(std::string numberAsString, bool requiredPositive) {
-	float parsedNumber = 0;
+int Expression::parseNumber(std::string numberAsString, bool requiredPositive) {
+	int parsedNumber = 0;
 
 	// remove leading and trailing spaces/unwanted characters
 	const char notLikedInNumbers[] = " \t\n\r\f\v";
@@ -134,7 +134,7 @@ float Expression::parseNumber(std::string numberAsString, bool requiredPositive)
 			std::string numberInRandomPar = numberAsString.substr(randomPos + 7, closeParentPos - randomPos - 7);
 
 			// make random number
-			parsedNumber = (float)(std::rand() % (int)parseNumber(numberInRandomPar, true));
+			parsedNumber = std::rand() % parseNumber(numberInRandomPar, true);
 		}
 		else {
 			throw std::invalid_argument("Missing ')'");
@@ -156,8 +156,8 @@ float Expression::parseNumber(std::string numberAsString, bool requiredPositive)
 	return parsedNumber;
 }
 
-float Expression::parseNumber(std::string number, bool requiredPositive, int maxValue) {
-	float parsedNumber = parseNumber(number, requiredPositive);
+int Expression::parseNumber(std::string number, bool requiredPositive, int maxValue) {
+	int parsedNumber = parseNumber(number, requiredPositive);
 
 	if (parsedNumber > maxValue) {
 		throw std::invalid_argument(std::string("Expression must be a number smaller than " + maxValue).c_str());
