+++ parser/parser/Expression.h
@@ -13,9 +13,9 @@ private:
 	static constexpr short variablesLength = 8;
 	int variables[variablesLength];
 
-	float expression(std::string, size_t = 0);
-	float term(std::string, size_t = 0);
-	float factor(std::string, size_t = 0);
+	int expression(std::string, size_t = 0);
+	int term(std::string, size_t = 0);
+	int factor(std::string, size_t = 0);
 public:
 	Expression(short padsCount) {
 		this->padsCount = padsCount;
@@ -23,9 +23,9 @@ public:
 			variables[i] = 0;
 		}
 	}
-	float parseExpression(std::string);
-	float parseNumber(std::string, bool = false);
-	float parseNumber(std::string, bool, int);
+	int parseExpression(std::string);
+	int parseNumber(std::string, bool = false);
+	int parseNumber(std::string, bool, int);
 
 	unsigned short getPadsCount() {
 		return this->padsCount;
